Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
anthropic==0.34.2
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
